requirements.txt
requests
python-dotenv
gradio
openai
anthropic
IPython
